+++ src/servers/model/utils/fetch-retry.ts
@@ -1,11 +1,14 @@
 /**
- * Copyright (c) 2018 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2018-2024 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author David Sehnal <david.sehnal@gmail.com>
+ * @author Adam Midlik <midlik@gmail.com>
  */
 
-import fetch, { Response } from 'node-fetch';
+import fetch, { AbortError, Response } from 'node-fetch';
 import { retryIf } from '../../../mol-util/retry-if';
+import { downloadGs } from '../../common/google-cloud-storage';
+
 
 const RETRIABLE_NETWORK_ERRORS = [
     'ECONNRESET', 'ENOTFOUND', 'ESOCKETTIMEDOUT', 'ETIMEDOUT',
@@ -20,7 +23,7 @@ export async function fetchRetry(url: string, timeout: number, retryCount: numbe
     const controller = new AbortController();
     const id = setTimeout(() => controller.abort(), timeout);
     const signal = controller.signal as any; // TODO: fix type
-    const result = await retryIf(() => fetch(url, { signal }), {
+    const result = await retryIf(() => wrapFetch(url, { signal }), {
         retryThenIf: r => r.status === 408 /** timeout */ || r.status === 429 /** too many requests */ || (r.status >= 500 && r.status < 600),
         // TODO test retryCatchIf
         retryCatchIf: e => isRetriableNetworkError(e),
@@ -30,4 +33,16 @@ export async function fetchRetry(url: string, timeout: number, retryCount: numbe
     clearTimeout(id);
 
     return result;
+}
+
+/** Like `fetch` but supports Google Cloud Storage (gs://) protocol. */
+export function wrapFetch(url: string, init?: fetch.RequestInit): Promise<Response> {
+    if (url.startsWith('gs://')) return fetchGS(url, init);
+    else return fetch(url, init);
+}
+
+async function fetchGS(url: string, init?: fetch.RequestInit): Promise<Response> {
+    if (init?.signal?.aborted) throw new AbortError('The user aborted a request.');
+    const data = await downloadGs(url);
+    return new Response(data, init);
 }